+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Bug Fixes / Nits
 - Fix HF LLM output error (#6737)
+- Add system message support for langchain message templates (#6743)
 
 ## [v0.7.1] - 2023-07-05